Move set_issue_status to gh_util
This is to avoid circular imports since some other modules need the
set_issue_status functionality.
