+++ marvin/gh_util.py
@@ -5,6 +5,15 @@ from typing import List
 
 from gidgethub.aiohttp import GitHubAPI
 
+# List of mutually exclusive status labels
+ISSUE_STATUS_LABELS = {
+    "needs_reviewer",
+    "awaiting_reviewer",
+    "awaiting_changes",
+    "needs_merger",
+    "awaiting_merger",
+}
+
 
 async def request_review(
     pull_url: str, gh_login: str, gh: GitHubAPI, token: str
@@ -61,3 +70,28 @@ async def get_installation_repositories(
         oauth_token=token,
     )
     return result["repositories"]
+
+
+async def set_issue_status(
+    issue: Dict[str, Any], status: str, gh: GitHubAPI, token: str
+) -> None:
+    """Sets the status of an issue while resetting other status labels"""
+    assert status in ISSUE_STATUS_LABELS
+
+    # depending on whether the issue is actually a pull request
+    issue_url = issue.get("issue_url", issue["url"])
+
+    # Labels are mutually exclusive, so clear other labels first.
+    labels = issue["labels"]
+    label_names = {label["name"] for label in labels}
+    # should never be more than one, but better to make it a set anyway
+    status_labels = label_names.intersection(ISSUE_STATUS_LABELS)
+    for label in status_labels:
+        if label == status:  # Don't touch the label we're supposed to set.
+            continue
+        await gh.delete(issue_url + "/labels/" + label, oauth_token=token)
+
+    if status not in status_labels:
+        await gh.post(
+            issue_url + "/labels", data={"labels": [status]}, oauth_token=token,
+        )
