+++ marvin/status.py
@@ -5,48 +5,15 @@ from gidgethub import routing
 from gidgethub import sansio
 from gidgethub.aiohttp import GitHubAPI
 
+from marvin import gh_util
 from marvin.command_router import CommandRouter
 
 router = routing.Router()
 command_router = CommandRouter()
 
-# List of mutually exclusive status labels
-ISSUE_STATUS_LABELS = {
-    "needs_reviewer",
-    "awaiting_reviewer",
-    "awaiting_changes",
-    "needs_merger",
-    "awaiting_merger",
-}
-
 NO_SELF_REVIEW_TEXT = f"""
 Sorry, you cannot set your own PR to `needs_merger` or `awaiting_merger`. Please wait for an external review. You may also actively search out a reviewer by pinging relevant people (look at the history of the files you're changing) or posting on discourse or IRC.
 """.strip()
-
-
-async def set_issue_status(
-    issue: Dict[str, Any], status: str, gh: GitHubAPI, token: str
-) -> None:
-    """Sets the status of an issue while resetting other status labels"""
-    assert status in ISSUE_STATUS_LABELS
-
-    # depending on whether the issue is actually a pull request
-    issue_url = issue.get("issue_url", issue["url"])
-
-    # Labels are mutually exclusive, so clear other labels first.
-    labels = issue["labels"]
-    label_names = {label["name"] for label in labels}
-    # should never be more than one, but better to make it a set anyway
-    status_labels = label_names.intersection(ISSUE_STATUS_LABELS)
-    for label in status_labels:
-        if label == status:  # Don't touch the label we're supposed to set.
-            continue
-        await gh.delete(issue_url + "/labels/" + label, oauth_token=token)
-
-    if status not in status_labels:
-        await gh.post(
-            issue_url + "/labels", data={"labels": [status]}, oauth_token=token,
-        )
 
 
 @router.register("pull_request", action="synchronize")
@@ -57,7 +24,7 @@ async def pull_request_synchronize(
     if "needs_merger" in {
         label["name"] for label in event.data["pull_request"]["labels"]
     }:
-        await set_issue_status(
+        await gh_util.set_issue_status(
             event.data["pull_request"], "awaiting_reviewer", gh, token
         )
 
@@ -80,12 +47,16 @@ async def issue_comment_event(
         if "awaiting_changes" in label_names:
             # A new comment by the author is probably some justification or request
             # for clarification. Action of the reviewer is needed.
-            await set_issue_status(event.data["issue"], "awaiting_reviewer", gh, token)
+            await gh_util.set_issue_status(
+                event.data["issue"], "awaiting_reviewer", gh, token
+            )
     else:
         # A new comment by somebody else is likely a review asking for
         # clarification or changes (provided it doesn't explicitly contain a
         # status command).
-        await set_issue_status(event.data["issue"], "awaiting_changes", gh, token)
+        await gh_util.set_issue_status(
+            event.data["issue"], "awaiting_changes", gh, token
+        )
 
 
 @router.register("pull_request_review", action="submitted")
@@ -95,7 +66,7 @@ async def pull_request_review_submitted_event(
     if len(command_router.find_commands(event.data["review"]["body"])) > 0:
         return
     if event.data["review"]["state"] == "changes_requested":
-        await set_issue_status(
+        await gh_util.set_issue_status(
             event.data["pull_request"], "awaiting_changes", gh, token
         )
 
@@ -104,21 +75,21 @@ async def pull_request_review_submitted_event(
 async def needs_reviewer_command(
     gh: GitHubAPI, token: str, issue: Dict[str, Any], **kwargs: Any
 ) -> None:
-    await set_issue_status(issue, "needs_reviewer", gh, token)
+    await gh_util.set_issue_status(issue, "needs_reviewer", gh, token)
 
 
 @command_router.register_command("/status awaiting_changes")
 async def awaiting_changes_command(
     gh: GitHubAPI, token: str, issue: Dict[str, Any], **kwargs: Any
 ) -> None:
-    await set_issue_status(issue, "awaiting_changes", gh, token)
+    await gh_util.set_issue_status(issue, "awaiting_changes", gh, token)
 
 
 @command_router.register_command("/status awaiting_reviewer")
 async def awaiting_reviewer_command(
     gh: GitHubAPI, token: str, issue: Dict[str, Any], **kwargs: Any
 ) -> None:
-    await set_issue_status(issue, "awaiting_reviewer", gh, token)
+    await gh_util.set_issue_status(issue, "awaiting_reviewer", gh, token)
 
 
 @command_router.register_command("/status needs_merger")
@@ -138,7 +109,7 @@ async def needs_merger_command(
             oauth_token=token,
         )
     else:
-        await set_issue_status(issue, "needs_merger", gh, token)
+        await gh_util.set_issue_status(issue, "needs_merger", gh, token)
 
 
 @command_router.register_command("/status awaiting_merger")
@@ -157,4 +128,4 @@ async def awaiting_merger_command(
             oauth_token=token,
         )
     else:
-        await set_issue_status(issue, "awaiting_merger", gh, token)
+        await gh_util.set_issue_status(issue, "awaiting_merger", gh, token)
